refactor: template_helper functions moved to Template impl block
It doesn't make much sense to have a Template struct and it functions in a
separate module
