+++ src/helpers/mod.rs
@@ -1,3 +1,2 @@
 pub mod file_helper;
 pub mod str_helper;
-pub mod template_helper;
