+++ src/lib.rs
@@ -1,12 +1,13 @@
 pub mod commands;
 pub mod constants;
 pub mod helpers;
+pub mod template;
 pub mod user_config;
 
 use std::env;
 
 use clap::Command;
-use commands::{new_command, init_command};
+use commands::{init_command, new_command};
 use constants::{CRATE_NAME, CRATE_VERSION};
 
 pub enum CreateableFileType {
@@ -23,12 +24,8 @@ pub fn run() {
 
     let base_cmd_args_matches = app.get_matches();
     let executed_cmd = match base_cmd_args_matches.subcommand() {
-        Some(("new", cmd_args)) => {
-            new_command::exec_command(cmd_args)
-        },
-        Some(("init", _)) => {
-            init_command::exec_command()
-        },
+        Some(("new", cmd_args)) => new_command::exec_command(cmd_args),
+        Some(("init", _)) => init_command::exec_command(),
         _ => Err(String::from("Unknown command")),
     };
